Escaping Newlines
Long statements can be broken up with the backslash ('\') character
before a newline. The newline will not be passed to the parser, and the
indentation of the following line will be ignored.
Block comments have been removed for ease of parsing. Now any code to
the right of two forward slashes ("//") is commented out. Indentation is
ignored on lines containing only comments.